Moved to unit test
